refactor: move TriggerKeyMapFromOtherAppsController to different package
